Force unsigned division (shift) in lower/upper_bound

+++ ualgo.h
@@ -368,7 +368,7 @@ ForwardIterator lower_bound (ForwardIterator first, ForwardIterator last, const 
 {
     ForwardIterator mid;
     while (first != last) {
-	mid = advance (first, distance (first,last) / 2);
+	mid = advance (first, size_t(distance (first,last)) / 2);
 	if (*mid < value)
 	    first = mid + 1;
 	else
@@ -396,7 +396,7 @@ ForwardIterator upper_bound (ForwardIterator first, ForwardIterator last, const 
 {
     ForwardIterator mid;
     while (first != last) {
-	mid = advance (first, distance (first,last) / 2);
+	mid = advance (first, size_t(distance (first,last)) / 2);
 	if (value < *mid)
 	    last = mid;
 	else
